# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/adm/emp": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет удалить работодателя из системы. Доступ имеют только пользователи роли ADMIN",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADMIN"
                ],
                "summary": "Удаление аккаунта работодателя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID работодателя, которого нужно удалить",
                        "name": "empID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус и краткую информацию ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
...
        "/emp": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет получить всю основную информацию про работодатля. Доступно всем авторизованным пользователям, но токен обязателен!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получить информцию про работодателя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID работодателя",
                        "name": "employeeID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!' и данные о работодателе",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.ResponseEmployeeInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса (токен или передача каких-либо других данных)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет обновить всю основную информацию о работодателе при помощи его персонального токена и тела запроса. Доступно только пользователям группы employee и ADMIN",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Обновить информцию о работодателе",
                "parameters": [
                    {
                        "description": "Данные о работодателе, на которые нужно обновить в системе",
                        "name": "Employee_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main_go_internal_api_Struct.RequestEmployee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!' и небольшую информацию",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.StatusInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса (токен или передача каких-либо других данных)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Позволяет добавлять нового работодателя в систему. В ответе клиент получит токен, с помощью которого сможет получить доступ к некоторому функционалу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Добавить нового работодателя",
                "parameters": [
                    {
                        "description": "Основные данные для добавления работодателя. В поле статус указывайте ID, который уже есть в системе!",
                        "name": "Employee_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main_go_internal_api_Struct.RequestEmployee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!', данные работодателя и новый токен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.ResponseCreateEmployee"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса (токен или передача каких-либо других данных)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/emp/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет получить всю основную информацию про всех работодатлей. Доступно только пользователям с ролью ADMIN",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получить информцию про всех работодателей",
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!' и массив всех данных о работодателях",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.SuccessEmployer"
                            }
                        }
                    },
                    "400": {
                        "description": "Возвращает ошибку, если не удалось получить данные из запроса (токен или передача каких-либо других данных)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "401": {
                        "description": "Возвращает ошибку, если у пользователя нету доступа к этому функционалу.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
        "/emp/auth": {
            "get": {
                "description": "Позволяет получить новый токен для работодателя, чтобы у него сохранился доступ к функционалу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Авторизовать работодателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email работодателя",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password работодателя",
                        "name": "Password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает статус 'Ok!', данные работодателя и новый токен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.ResponseCreateEmployee"
                            }
                        }
                    },
                    "500": {
                        "description": "Возвращает ошибку, если на сервере произошла непредвиденная ошибка.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main_go_internal_api_Struct.InfoError"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main_go_internal_api_Struct.InfoError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main_go_internal_api_Struct.Ok": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "main_go_internal_api_Struct.RequestEmployee": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "inn": {
                    "type": "string"
                },
                "name_organization": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
...
        "main_go_internal_api_Struct.ResponseCreateEmployee": {
            "type": "object",
            "properties": {
                "candidate_Info": {
                    "$ref": "#/definitions/main_go_internal_api_Struct.SuccessEmployer"
                },
                "status": {
                    "$ref": "#/definitions/main_go_internal_api_Struct.Ok"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "main_go_internal_api_Struct.ResponseEmployeeInfo": {
            "type": "object",
            "properties": {
                "employee": {
                    "$ref": "#/definitions/main_go_internal_api_Struct.SuccessEmployer"
                },
                "status": {
                    "$ref": "#/definitions/main_go_internal_api_Struct.Ok"
                }
            }
        },
...
        "main_go_internal_api_Struct.StatusInfo": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "main_go_internal_api_Struct.SuccessEmployer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inn": {
                    "type": "string"
                },
                "nameOrganization": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "object",
                    "properties": {
                        "createdAt": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}